Working libwavernn high speed multicore inference
