+++ vocoder/libwavernn/inference.py
@@ -1,11 +1,12 @@
 import math
 import time
+
+from concurrent import futures
 from concurrent.futures import ThreadPoolExecutor
 
 from config.hparams import sp, wavernn_fatchord, wavernn_geneing, wavernn_runtimeracer
 from vocoder.models import base
 from vocoder.audio import decode_mu_law, de_emphasis
-from vocoder.utils import folding
 from pathlib import Path
 import numpy as np
 import torch
@@ -75,6 +76,7 @@ class Vocoder:
 
         # Adjust mel range range to [-1, 1] (normalization)
         mel = mel.T.astype(np.float32) / sp.max_abs_value
+        wave_len = mel.shape[1] * sp.hop_size
 
         output = None
         wrapper_count = len(self._processing_thread_wrappers)
@@ -92,16 +94,12 @@ class Vocoder:
             min_target = hp_wavernn.gen_target
             min_overlap = hp_wavernn.gen_overlap
 
-            _, mel_size, _ = mel.size()
-            optimal_target = math.ceil(((mel_size - min_overlap) / wrapper_count) - min_overlap)
+            optimal_target = math.ceil(((wave_len - min_overlap) / wrapper_count) - min_overlap)
             if optimal_target < min_target:
                 optimal_target = min_target
 
             # Do the folding
-            mels = mel
-            with torch.no_grad():
-                wave_len = (mels.size(-1) - 1) * sp.hop_size
-                mels = folding.fold_with_overlap(mels, optimal_target, min_overlap)
+            mels = self.fold_mel_with_overlap(mel, optimal_target, min_overlap)
 
             # Render using multithreading
             chunks_count = len(mels)
@@ -110,28 +108,12 @@ class Vocoder:
                 for tID in range(chunks_count):
                     threads.append(executor.submit(self.vocode_thread, tID, mels[tID]))
 
-            # wait for processing done
-            processing_done = False
-            while not processing_done:
-                done_cnt = 0
-                for t in threads:
-                    if t.done():
-                        done_cnt += 1
-                if done_cnt == len(threads):
-                    processing_done = True
-                else:
-                    time.sleep(0.01)
-
             # Unfold the results
             output = []
             for t in threads:
-                output.append(t.result().numpy())
-
-            with torch.no_grad():
-                output = torch.stack(output).transpose(0, 1)
-                output = output.cpu().numpy()
-                output = output.astype(np.float64)
-                output = folding.xfade_and_unfold(output, optimal_target, min_overlap)
+                output.append(t.result())
+            output = np.stack(output, axis=0)
+            output = self.unfold_wav_with_overlap(output, optimal_target, min_overlap)
 
         if hp_wavernn.mu_law:
             # Do MuLaw decode over the whole generated audio for optimal normalization
@@ -149,3 +131,67 @@ class Vocoder:
     def vocode_thread(self, tID, chunk):
         return self._processing_thread_wrappers[tID].melToWav(chunk)
 
+    def fold_mel_with_overlap(self, mel, target, overlap):
+        # This folding happens before upsampling on the raw mel
+        mel_len = mel.shape[1]
+        mel_target = math.ceil(target / sp.hop_size)
+        mel_overlap = math.ceil(overlap / sp.hop_size)
+
+        # Calculate variables needed
+        num_folds = (mel_len - mel_overlap) // (mel_target + mel_overlap)
+        extended_len = num_folds * (mel_overlap + mel_target) + mel_overlap
+        remaining = mel_len - extended_len
+
+        # Pad if some time steps poking out
+        if remaining != 0:
+            num_folds += 1
+            padding = mel_target + 2 * mel_overlap - remaining
+            total = mel_len + padding
+            # Apply padding to the mel
+            padded_mel = np.zeros(shape=(mel.shape[0], total), dtype=np.float32)
+            padded_mel[:, :mel_len] = mel
+            mel = padded_mel
+
+        # Get the values for the folded mel parts
+        folded = []
+        for i in range(num_folds):
+            start = i * (mel_target + mel_overlap)
+            end = start + mel_target + 2 * mel_overlap
+            folded_elem = mel[:, start:end]
+            folded.append(folded_elem)
+
+        return folded
+
+    def unfold_wav_with_overlap(self, wav, target, overlap):
+
+        num_folds, length = wav.shape
+        target = length - 2 * overlap
+        total_len = num_folds * (target + overlap) + overlap
+
+        # Need some silence for the rnn warmup
+        silence_len = overlap // 2
+        fade_len = overlap - silence_len
+        silence = np.zeros((silence_len), dtype=np.float64)
+
+        # Equal power crossfade
+        t = np.linspace(-1, 1, fade_len, dtype=np.float64)
+        fade_in = np.sqrt(0.5 * (1 + t))
+        fade_out = np.sqrt(0.5 * (1 - t))
+
+        # Concat the silence to the fades
+        fade_in = np.concatenate([silence, fade_in])
+        fade_out = np.concatenate([fade_out, silence])
+
+        # Apply the gain to the overlap samples
+        wav[:, :overlap] *= fade_in
+        wav[:, -overlap:] *= fade_out
+
+        unfolded = np.zeros((total_len), dtype=np.float64)
+
+        # Loop to add up all the samples
+        for i in range(num_folds):
+            start = i * (target + overlap)
+            end = start + target + 2 * overlap
+            unfolded[start:end] += wav[i]
+
+        return unfolded
